requirements.txt
aiohttp==3.9.1
anthropic==0.18.1
openai==1.12.0
ollama==0.1.7
streamlit==1.31.0
plotly==5.17.0
pandas==2.2.0
numpy==1.26.3
scikit-learn==1.4.0
pydantic==2.5.3
python-dotenv==1.0.1
rich==13.7.0
asyncio-throttle==1.0.2
jinja2==3.1.3
pytest==8.2.0
pytest-asyncio==0.24.0
black==24.1.1
mypy==1.8.0
docker==7.0.0